refactored for new PRs that were merged
